make a post a get - simplicity
